Added Methods Description

+++ reelsplayer/src/main/java/com/shahid/iqbal/reelsplayer/configs/ReelsConfig.kt
@@ -11,6 +11,19 @@ import com.shahid.iqbal.reelsplayer.actions.VideoScalingMode
 import com.shahid.iqbal.reelsplayer.components.DefaultVideoLoader
 
 
+/**
+ * Configuration class for the Reels feature.
+ *
+ * @property playerResizeMode The resize mode for the video player.
+ * @property videoScalingMode The scaling mode for video display.
+ * @property repeatMode The repeat mode for video playback.
+ * @property thumbnailDisplayMode The display mode for thumbnails.
+ * @property showControlsMenu Whether to show the controls menu.
+ * @property playerSize The size of the video player.
+ * @property reelDetail A composable function that takes an integer representing the video index
+ *                      and displays details for the specified reel.
+ * @property playerLoader A composable function to display the video loader.
+ */
 @Keep
 data class ReelsConfig(
     var playerResizeMode: PlayerResizeMode = PlayerResizeMode.FILL,
@@ -19,6 +32,7 @@ data class ReelsConfig(
     var thumbnailDisplayMode: ThumbnailDisplayMode = ThumbnailDisplayMode.FILL,
     val showControlsMenu: Boolean = false,
     var playerSize: Size = Size(LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT),
-    var reelDetail: (@Composable () -> Unit)? = null,
+    var reelDetail: (@Composable (Int) -> Unit)? = null,
     var playerLoader: (@Composable () -> Unit)? = { DefaultVideoLoader() },
 )
+
